backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-jose
passlib[bcrypt]
bcrypt==3.2.0
python-dotenv
sqlalchemy
pydantic
